Include user ID in details saved to Cognito shopping list dataset
This is a workaround to pass the relevant user ID to the Lambda function
that handles Cognito sync events. Sync events only include the dataset(s)
that were modified, so the function has no way to know which Alexa user
to specify when sending an update message to the Alexa Skill.